orders/src/events/listeners/__test__/expiration-complete-listener.test.ts
import mongoose from "mongoose";
import { Ticket } from "../../../models/ticket";
import { natsWrapper } from "../../../nats-wrapper";
import ExpirationCompleteListener from "../expiration-complete-listener";
import { IExpirationCompleteEvent, OrderStatus } from "@ticketing-nir/common";
import { Order } from "../../../models/order";
import { Message } from "node-nats-streaming";

const setup = async () => {
  const listener = new ExpirationCompleteListener(natsWrapper.client);

  const ticket = Ticket.build({ title: "Concert", price: 20, id: new mongoose.Types.ObjectId().toHexString() });
  ticket.set({ orderId: new mongoose.Types.ObjectId().toHexString() });
  await ticket.save();

  const order = Order.build({
    ticket: ticket,
    expiresAt: new Date(Date.now() + 1000 * 60 * 6),
    status: OrderStatus.Created,
    userId: "sdfsdf",
  });
  await order.save();

  const data: IExpirationCompleteEvent["data"] = {
    orderId: order.id,
  };
  // @ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, order, data, msg, ticket };
};

it("updates the order status to Cancelled", async () => {
  const { data, listener, msg, order } = await setup();

  await listener.onMessage(data, msg);

  const updatedOrder = await Order.findById(order.id);
  expect(updatedOrder?.status).toEqual(OrderStatus.Cancelled);
});

it("emit an orderCancelled event", async () => {
  const { data, listener, msg, order } = await setup();

  await listener.onMessage(data, msg);
  expect(natsWrapper.client.publish).toHaveBeenCalled();
  const eventData = JSON.parse((natsWrapper.client.publish as jest.Mock).mock.calls[0][1]);
  expect(eventData.id).toEqual(order.id);
});

it("Acks the message", async () => {
  const { data, listener, msg } = await setup();
  // Call the onMessage function with the data object + message object
  await listener.onMessage(data, msg);
  // Write assertions to make sure a ack was called
  expect(msg.ack).toHaveBeenCalled();
});
